Reorganize tests to reflect pandas structure
