+++ tests/snippets/frame/test_typing.py
