+++ tests/snippets/scalar/interval/test_typing.py
