+++ tests/snippets/scalar/timestamp/test_typing.py
