+++ tests/snippets/series/test_typing.py
